warning removed

+++ lib/model/sendOrder.dart
@@ -1,6 +1,7 @@
 import 'dart:convert';
 
-SendOrderModel sendOrderModelFromJson(String str) => SendOrderModel.fromJson(json.decode(str));
+SendOrderModel sendOrderModelFromJson(String str) =>
+    SendOrderModel.fromJson(json.decode(str));
 
 String sendOrderModelToJson(SendOrderModel data) => json.encode(data.toJson());
 
@@ -24,7 +25,7 @@ class SendOrderModel {
   int addressId;
   double tax;
   double deliveryCharge;
-  int coupon;
+  double coupon;
   int voucher;
   int offer;
   String shopCategory;
@@ -34,34 +35,34 @@ class SendOrderModel {
   List<Item> items;
 
   factory SendOrderModel.fromJson(Map<String, dynamic> json) => SendOrderModel(
-    userId: json["user_id"],
-    addressId: json["address_id"],
-    tax: json["tax"].toDouble(),
-    deliveryCharge: json["delivery_charge"].toDouble(),
-    coupon: json["coupon"],
-    voucher: json["voucher"],
-    offer: json["offer"],
-    shopCategory: json["shop_category"],
-    paymentOption: json["payment_option"],
-    lng: json["lng"],
-    lat: json["lat"],
-    items: List<Item>.from(json["items"].map((x) => Item.fromJson(x))),
-  );
+        userId: json["user_id"],
+        addressId: json["address_id"],
+        tax: json["tax"].toDouble(),
+        deliveryCharge: json["delivery_charge"].toDouble(),
+        coupon: json["coupon"],
+        voucher: json["voucher"],
+        offer: json["offer"],
+        shopCategory: json["shop_category"],
+        paymentOption: json["payment_option"],
+        lng: json["lng"],
+        lat: json["lat"],
+        items: List<Item>.from(json["items"].map((x) => Item.fromJson(x))),
+      );
 
   Map<String, dynamic> toJson() => {
-    "user_id": userId,
-    "address_id": addressId,
-    "tax": tax,
-    "delivery_charge": deliveryCharge,
-    "coupon": coupon,
-    "voucher": voucher,
-    "offer": offer,
-    "shop_category": shopCategory,
-    "payment_option": paymentOption,
-    "lng": lng,
-    "lat": lat,
-    "items": List<dynamic>.from(items.map((x) => x.toJson())),
-  };
+        "user_id": userId,
+        "address_id": addressId,
+        "tax": tax,
+        "delivery_charge": deliveryCharge,
+        "coupon": coupon,
+        "voucher": voucher,
+        "offer": offer,
+        "shop_category": shopCategory,
+        "payment_option": paymentOption,
+        "lng": lng,
+        "lat": lat,
+        "items": List<dynamic>.from(items.map((x) => x.toJson())),
+      };
 }
 
 class Item {
@@ -72,18 +73,18 @@ class Item {
   });
 
   int productId;
-  double  price;
+  double price;
   int qty;
 
   factory Item.fromJson(Map<String, dynamic> json) => Item(
-    productId: json["product_id"],
-    price: json["price"],
-    qty: json["qty"],
-  );
+        productId: json["product_id"],
+        price: json["price"],
+        qty: json["qty"],
+      );
 
   Map<String, dynamic> toJson() => {
-    "product_id": productId,
-    "price": price,
-    "qty": qty,
-  };
+        "product_id": productId,
+        "price": price,
+        "qty": qty,
+      };
 }
